Compilation fixes for x86_64.
Added quite a few missing components to image:

alert arp bc CortexAddOnHost

CDPlayer CharacterMap DeskCalc
Devices DiskUsage Expander LaunchBox MediaConverter NetworkStatus
PoorMan Pulse Screenshot ShowImage SoundRecorder TV Workspaces

DataTranslations Media Printers ScreenSaver

BSnow Cortex FontDemo OverlayImage Pairs

cddb_daemon midi_server power_daemon print_server print_addon_server

Canon LIPS3 Compatible Canon LIPS4 Compatible Gutenprint
PCL5 Compatible PCL6 Compatible PDF Writer PS Compatible

libgame.so libmedia.so

+++ src/add-ons/print/drivers/pdf/source/DocInfoWindow.cpp
@@ -425,7 +425,7 @@ void
 DocInfoWindow::_EmptyKeyList()
 {
 	while (fKeyList->CountItems() > 0L)
-		delete fKeyList->RemoveItem(0L);
+		delete fKeyList->RemoveItem((int32)0);
 }
